+++ CHANGELOG.md
@@ -4,6 +4,10 @@ follows the formatting recommendations from [Keep a
 CHANGELOG](http://keepachangelog.com/). This project adheres to [Semantic
 Versioning](http://semver.org/).
 
+## [Unreleased][unreleased]
+### Added
+- `split` for combinining `match` and `sep`.
+
 ## [0.2] - 2015-4-5
 ### Changed
 - Replace basic routing (monoidal matching) with lens routing.
@@ -22,7 +26,7 @@ Versioning](http://semver.org/).
 - HTTP parameters handling for both *query string* and form encoded `POST`s.
 - Trailing slash redirection.
 
-
+[unreleased]: https://github.com/jdnavarro/nero/compare/v0.2...HEAD
 [0.2]: https://github.com/jdnavarro/nero/compare/v0.1.1...v0.2
 [0.1.1]: https://github.com/jdnavarro/nero/compare/v0.1...v0.1.1
 [0.1]: https://github.com/jdnavarro/nero/compare/a2c3f720...v0.1